requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
python-multipart==0.0.9          # dependência do FastAPI p/ formulários
email-validator==2.1.1           # dependência interna do FastAPI
python-dotenv==1.0.1             # NOVO
pydantic-settings==2.2.1         # NOVO
